Updating unit tests, close #922

+++ Tests.Unit/Services/LogServiceTests.cs
@@ -3,8 +3,10 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using System.Threading;
+using System.Threading.Tasks;
 using EmbyStat.Common.Enums;
 using EmbyStat.Common.Extensions;
+using EmbyStat.Common.Models.Entities;
 using EmbyStat.Common.Models.Settings;
 using EmbyStat.Services;
 using EmbyStat.Services.Interfaces;
@@ -40,7 +42,9 @@ namespace Tests.Unit.Services
             _settingsServiceMock.Setup(x => x.GetAppSettings()).Returns(settings);
             _settingsServiceMock.Setup(x => x.GetUserSettings()).Returns(userSettings);
 
-            var service = new LogService(_settingsServiceMock.Object);
+            var embyServiceMock = new Mock<IEmbyService>();
+
+            var service = new LogService(_settingsServiceMock.Object, embyServiceMock.Object);
 
             File.Create(Path.Combine("config", "Logs-test1", "log1.txt").GetLocalPath());
             Thread.Sleep(1001);
@@ -66,11 +70,12 @@ namespace Tests.Unit.Services
             var settings = new AppSettings { Dirs = new Dirs { Logs = "Logs-test2", Config = "config" } };
             _settingsServiceMock.Setup(x => x.GetAppSettings()).Returns(settings);
 
-            var userSettings = new UserSettings();
-            userSettings.KeepLogsCount = 1;
+            var userSettings = new UserSettings { KeepLogsCount = 1 };
 
             _settingsServiceMock.Setup(x => x.GetUserSettings()).Returns(userSettings);
-            var service = new LogService(_settingsServiceMock.Object);
+            var embyServiceMock = new Mock<IEmbyService>();
+
+            var service = new LogService(_settingsServiceMock.Object, embyServiceMock.Object);
 
             File.Create(Path.Combine("config", "Logs-test2", "log1.txt").GetLocalPath());
             Thread.Sleep(1001);
@@ -83,54 +88,67 @@ namespace Tests.Unit.Services
         }
 
         [Fact]
-        public void GetLogStream()
+        public async void GetLogStream()
         {
+            if (Directory.Exists(Path.Combine("config", "Logs-test3").GetLocalPath()))
+            {
+                Directory.Delete(Path.Combine("config", "Logs-test3").GetLocalPath(), true);
+            }
+
             Directory.CreateDirectory(Path.Combine("config", "Logs-test3").GetLocalPath());
 
             var settings = new AppSettings { Dirs = new Dirs { Logs = "Logs-test3", Config = "config" } };
             var userSettings = new UserSettings { Emby = new EmbySettings { UserName = "reggi", ServerProtocol = ConnectionProtocol.Http, ServerAddress = "192.168.1.1", ServerPort = 8001 }, Tvdb = new TvdbSettings { ApiKey = "0000" } };
             _settingsServiceMock.Setup(x => x.GetAppSettings()).Returns(settings);
             _settingsServiceMock.Setup(x => x.GetUserSettings()).Returns(userSettings);
-            var service = new LogService(_settingsServiceMock.Object);
+
+            var embyServiceMock = new Mock<IEmbyService>();
+            embyServiceMock.Setup(x => x.GetServerInfo()).Returns(Task.FromResult(new ServerInfo { Id = Guid.NewGuid().ToString() }));
+
+            var service = new LogService(_settingsServiceMock.Object, embyServiceMock.Object);
 
             var line = "Log line: http://192.168.1.1:8001; ApiKey:0000; username:reggi";
             File.AppendAllText(Path.Combine("config", "Logs-test3", "log1.txt").GetLocalPath(), line);
 
-            var stream = service.GetLogStream("log1.txt", false);
+            var stream = await service.GetLogStream("log1.txt", false);
             using (var reader = new StreamReader(stream, Encoding.UTF8))
             {
                 var lines = reader.ReadToEnd();
                 lines.Should().Be(line);
             }
-
-            Directory.Delete(Path.Combine("config", "Logs-test3").GetLocalPath(), true);
         }
 
         [Fact]
-        public void GetAnonymousLogStream()
+        public async void GetAnonymousLogStream()
         {
+            if (Directory.Exists(Path.Combine("config", "Logs-test4").GetLocalPath()))
+            {
+                Directory.Delete(Path.Combine("config", "Logs-test4").GetLocalPath(), true);
+            }
+
             Directory.CreateDirectory(Path.Combine("config", "Logs-test4").GetLocalPath());
 
             var settings = new AppSettings { Dirs = new Dirs { Logs = "Logs-test4", Config = "config" } };
-            var userSettings = new UserSettings { Emby = new EmbySettings { UserName = "reggi", ServerProtocol = ConnectionProtocol.Http, ServerAddress = "192.168.1.1", ServerPort = 8001 }, Tvdb = new TvdbSettings { ApiKey = "0000" } };
+            var userSettings = new UserSettings { Emby = new EmbySettings { UserName = "reggi", ServerProtocol = ConnectionProtocol.Http, ServerAddress = "192.168.1.1", ServerPort = 8001, AccessToken = "123456" }, Tvdb = new TvdbSettings { ApiKey = "0000" } };
             _settingsServiceMock.Setup(x => x.GetAppSettings()).Returns(settings);
             _settingsServiceMock.Setup(x => x.GetUserSettings()).Returns(userSettings);
-            var service = new LogService(_settingsServiceMock.Object);
 
-            var line = "Log line: http://192.168.1.1:8001; ApiKey:0000; username:reggi\r\n";
+            var embyServiceMock = new Mock<IEmbyService>();
+            embyServiceMock.Setup(x => x.GetServerInfo()).Returns(Task.FromResult(new ServerInfo { Id = "654321" }));
+
+            var service = new LogService(_settingsServiceMock.Object, embyServiceMock.Object);
+
+            const string line = "Log line: http://192.168.1.1:8001; ApiKey:0000; username:reggi; ws://192.168.1.1:8001; accessToken:123456; serverId:654321\r\n";
+            const string anonymousLine = "Log line: http://xxx.xxx.xxx.xxx:xxxx; ApiKey:xxxxxxxxxxxxxx; username:{EMBY ADMIN USER}; wss://xxx.xxx.xxx.xxx:xxxx; accessToken:xxxxxxxxxxxxxx; serverId:xxxxxxxxxxxxxx\r\n";
             File.AppendAllText(Path.Combine("config", "Logs-test4", "log1.txt").GetLocalPath(), line);
 
-            var stream = service.GetLogStream("log1.txt", true);
+
+            var stream = await service.GetLogStream("log1.txt", true);
             using (var reader = new StreamReader(stream))
             {
                 var lines = reader.ReadToEnd();
-                line = line.Replace(userSettings.FullEmbyServerAddress, "http://xxx.xxx.xxx.xxx:xxxx");
-                line = line.Replace(userSettings.Tvdb.ApiKey, "xxxxxxxxxxxxxx");
-                line = line.Replace(userSettings.Emby.UserName, "{EMBY ADMIN USER}");
-                lines.Should().Be(line);
+                lines.Should().Be(anonymousLine);
             }
-
-            Directory.Delete(Path.Combine("config", "Logs-test4").GetLocalPath(), true);
         }
     }
 }
